Fix broken import on localized CRM
As it turned out the localization could break the import process. The
CRM_Contribute_BAO_Contribution::buildOptions function returns the options with localized labels.
Due to this, the contribution_status_id is mapped to the default invalid value, that blocked the
contribution creation. To handle this issue, the localized version of the label has to be used for mapping.