Use hardcoded contribution_status_ids
As the labels of the status ids could be on fully custom value, the prediction logic
could return the fallback 0 values, so that the contribution will be cancelled.
To prevent this issue, the ids now wired to the function. The ids were used from a
fresh install of the CRM.